moving the search bar
