+++ src/App.jsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useState } from 'react'
 import Header from './components/Header/Header';
 import Router from './Router';
 import "../src/assets/scss/styles.scss"
@@ -6,7 +6,7 @@ import AppContext from './context/AppContext';
 import { Footer } from './components/Footer';
 
 function App() {
-  const [open, setOpen] = React.useState(false);
+  const [open, setOpen] = useState(false);
 
   const handleClick = () => {
     setOpen(true);
