+++ src/assets/scss/_components.scss
@@ -144,11 +144,6 @@
   color: cadetblue;
   font-size: 14px;
 }
-.search__field {
-  display: flex ;
-  align-items: center ;
-  margin: 10px 16px;
-}
 
 .c-cart-list-item, .c-favorite-list-item {
   display: flex;
